// frontend/src/generated/types/sysAccessLogMaster.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface SysAccessLogMaster {
  sysId?: string;
  userId?: string;
  userNm?: string;
  ipAddress?: string;
  loginDatetime?: string;
  logoutDatetime?: string;
}